Constants.class moved to Factory package
