--- src/main/java/ua/goit/dao/Constants.java
@@ -1,8 +1,0 @@
-package ua.goit.dao;
-public class Constants {
-  static final String URL_TO_MYSQLSERVER = "jdbc:mysql://localhost/";
-  static final String DB_NAME = "kickstarter";
-  static final String DB_URL = URL_TO_MYSQLSERVER + DB_NAME;
-  static final String USER = "kickstarter";
-  static final String PASS = "kickstarter";
-}
