+++ src/main/java/ua/goit/factory/Constants.java
@@ -1,0 +1,9 @@
+package ua.goit.factory;
+public class Constants {
+	static final String URL_TO_MYSQLSERVER = "jdbc:mysql://localhost/";
+	static final String DB_NAME = "kickstarter";
+	static final String DB_URL = URL_TO_MYSQLSERVER + DB_NAME;
+	static final String USER = "kickstarter";
+	static final String PASS = "kickstarter";
+	static final int MAX_CONN = 2;
+}
