+++ src/main/java/ua/goit/factory/PrintOutCategories.java
@@ -21,36 +21,16 @@ public class PrintOutCategories extends HttpServlet {
 		connMgr = DBConnectionManager.getInstance();
 	}
 
-
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
 		response.setContentType("text/html");	
-		Connection con = connMgr.getConnection("idb");
+		Connection con = connMgr.getConnection();
 		if (con == null) {
 			response.getWriter().print("Can't get connection");						
 		}
 		else {
 			response.getWriter().print("Connection is done");
 		}
-
-
-
-		/*Set<String> categories = new Categories().getCategories();
-
-				//String jspUrl = request.getContextPath() + "/Projects.jsp";
-				RequestDispatcher dispatcher = request.getRequestDispatcher("/Projects");
-				request.setAttribute("set", categories);
-				dispatcher.forward(request, response);*/
-		//
-		//		response.getWriter().print("<html>");
-		//		response.getWriter().print("<body>");
-		//		for (String i : categories) {
-		//			response.getWriter().println("<a href=http://localhost:8080/group1/Projects?category=" 
-		//					+ i + ">" + i + "</a><br>");	
-		//		}
-		//		response.getWriter().print("</body>");
-		//		response.getWriter().print("</html>");
-		
-		connMgr.freeConnection("idb", con);
+		connMgr.freeConnection(con);
 	}
 
 	public void destroy() {
