--- src/main/resources/db.properties
@@ -1,6 +1,0 @@
-drivers=com.mysql.jdbc.Driver
-
-idb.url=jdbc:mysql://localhost:3306/employees
-idb.maxconn=2
-idb.user=root
-idb.password=1qazse4
